requirements.txt
psutil>=5.9.0
py-cpuinfo>=9.0.0
GPUtil>=1.4.0
tabulate>=0.9.0
wmi>=1.5.1; platform_system == "Windows"
requests>=2.28.0
rich>=13.0.0
netifaces>=0.11.0
speedtest-cli>=2.1.3 